refactor: rename folder as well
